+++ Tests/AwesomeAssertions.Specs/Primitives/StringAssertionSpecs.Be.cs
@@ -166,7 +166,14 @@ public partial class StringAssertionSpecs
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Expected string to be \"ABC \" because I say so, but it misses some extra whitespace at the end.");
+                """
+                *index 3*
+                      ↓ (actual)
+                  "ABC"
+                  "ABC "
+                      ↑ (expected).
+                """
+            );
         }
 
         [Fact]
@@ -178,7 +185,14 @@ public partial class StringAssertionSpecs
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Expected string to be \"ABC\" because I say so, but it has unexpected whitespace at the end.");
+                """
+                *index 3*
+                      ↓ (actual)
+                  "ABC "
+                  "ABC"
+                      ↑ (expected).
+                """
+            );
         }
 
         [Fact]
@@ -387,9 +401,32 @@ public partial class StringAssertionSpecs
             // Assert
             act.Should().Throw<XunitException>()
                 .Which.Message.Should().Match("""
-                    *"…class Foo { }"
-                    *"…class Foo { };"
-                    *(expected).
+                     *"…class Foo { }"
+                     *"…class Foo { };"
+                     *(expected).
+                     """);
+        }
+
+        [Fact]
+        public void Prefix_before_mismatch_is_truncated_earlier_than_suffix()
+        {
+            const string subject = "diff the text with very long suffix after point where mismatch is found";
+            const string expected = "diff the texT with very long suffix after point where mismatch is found";
+
+            // Act
+            Action act = () => subject.Should().Be(expected);
+
+            // Assert
+            act
+                .Should()
+                .Throw<XunitException>()
+                .WithMessage(
+                    """
+                    *index 12*
+                               ↓ (actual)
+                      "…the text with very long suffix after point where mismatch is…"
+                      "…the texT with very long suffix after point where mismatch is…"
+                               ↑ (expected).
                     """);
         }
     }